refactor: Split Game.cs and Board.cs into logical partial classes
Game.cs (2185 lines) split into 6 partial files:
- Game.Core.cs: initialization, properties, state
- Game.Movement.cs: MovePiece and movement validation
- Game.Abilities.cs: on-stop and passive ability implementations
- Game.TurnPhases.cs: turn/phase management
- Game.Scoring.cs: score tracking and coin events
- Game.Helpers.cs: charge, ice patches, board queries

Board.cs (551 lines) split into 3 partial files:
- Board.Core.cs: initialization, tile access, bounds
- Board.Obstacles.cs: rocks, lakes, fences, ice patches
- Board.Passability.cs: passability checks, movement validation

No behavior changes — purely organizational. All tests passing.
